# refactor to optimize dnsmessage function