… .golangci.yml
… docs/dnsjson.md