Optimize the `Flatten` function of the DNSMessage to reduce CPU usage.

Before
```
./dnsutils$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/dmachard/go-dnscollector/dnsutils
cpu: Intel(R) Core(TM) i5-7200U CPU @ 2.50GHz
BenchmarkDnsMessage_ToDNSTap-4                    836766              1229 ns/op
BenchmarkDnsMessage_ToExtendedDNSTap-4            520940              2116 ns/op
BenchmarkDnsMessage_ToJSON-4              167796              6910 ns/op
BenchmarkDnsMessage_ToFlatJSON-4            6794            172489 ns/op
BenchmarkDnsMessage_ToTextFormat-4                732069              1457 ns/op
BenchmarkDnsMessage_ToPacketLayer-4              1171467               969.5 ns/op
PASS
ok      github.com/dmachard/go-dnscollector/dnsutils    2.056s
```

After refactor with default configuration:

```
./dnsutils$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/dmachard/go-dnscollector/dnsutils
cpu: Intel(R) Core(TM) i5-7200U CPU @ 2.50GHz
BenchmarkDnsMessage_ToDNSTap-4                    993242              1130 ns/op
BenchmarkDnsMessage_ToExtendedDNSTap-4            618400              1951 ns/op
BenchmarkDnsMessage_ToJSON-4                      190939              6584 ns/op
BenchmarkDnsMessage_ToFlatJSON-4                   19868             55533 ns/op
BenchmarkDnsMessage_ToTextFormat-4               2600718               460.7 ns/op
BenchmarkDnsMessage_ToPacketLayer-4              1171467               969.5 ns/op
PASS
ok      github.com/dmachard/go-dnscollector/dnsutils    1.810s
```